# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Platform Chaos Reliability Hub API
  version: 0.2.0
  description: REST API for chaos experiment catalog, service risk, approval requests, runbooks, integrations and incident summaries.
servers:
  - url: http://localhost:8080
paths:
...
  /api/approvals/requests:
    get:
      summary: List approval requests
      responses:
        "200":
          description: Approval requests
    post:
      summary: Create approval request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalRequestInput"
      responses:
        "201":
          description: Approval request created
        "400":
          description: Missing required guardrails
...
  /api/ai/summarize:
    post:
      summary: Generate deterministic incident summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AISummaryInput"
      responses:
        "200":
          description: Incident summary
...
components:
...
  schemas:
    ApprovalRequestInput:
      type: object
      required: [experimentId, serviceId, requestedBy, riskScore]
      properties:
        experimentId:
          type: string
        serviceId:
          type: string
        requestedBy:
          type: string
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
        riskLevel:
          type: string
        blastRadius:
          type: string
        rollbackThreshold:
          type: string
        runbookId:
          type: string
        evidenceLinks:
          type: array
          items:
            type: string
...
    AISummaryInput:
      type: object
      required: [experimentName, affectedService, riskScore]
      properties:
        experimentName:
          type: string
        affectedService:
          type: string
        metricsSnapshot:
          type: object
          additionalProperties:
            type: string
        logsOrEvents:
          type: array
          items:
            type: string
        riskScore:
          type: integer